+++ src/vrl/vrl_dns.c
@@ -442,4 +442,6 @@ void vrl_host_builtins_init(void)
 	vrl_register("dns_lookup", fn_dns_lookup);
 	vrl_register("reverse_dns", fn_reverse_dns);
 	vrl_http_builtins_init();
+	vrl_enrich_builtins_init();
+	vrl_secret_builtins_init();
 }